Disable cpu info metric on non x86 arches
On an IBM Z system it was observed that the parsecpuinfo function was crashing because the parsing of the CPU
info is specific to x86 today - https://bugzilla.redhat.com/show_bug.cgi?id=1789260. For now, the short
term fix is to disable cpu info metrics on non x86 arches and in the meanwhile we can work on the fix for
supporting cpu info metric reporting for multi-arches.